backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
# async Postgres
databases[postgresql]
asyncpg
sqlalchemy
psycopg[binary]
pydantic
aiofiles
pytest
